--- app/processing/payment/calculate-payment-schedule/index.js
@@ -1,5 +1,0 @@
-const calculatePaymentSchedule = require('./calculate-payment-schedule')
-
-module.exports = {
-  calculatePaymentSchedule
-}
