+++ app/processing/payment/calculate-scheduled-payments/get-period.js
