+++ app/processing/payment/calculate-scheduled-payments/index.js
@@ -1,0 +1,5 @@
+const calculateScheduledPayments = require('./calculate-payment-schedule')
+
+module.exports = {
+  calculateScheduledPayments
+}
